# services/auth/api/docs/swagger.yaml
definitions:
  github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.AddPermissionScopeRequest:
    properties:
      client_platform_id:
        type: string
      method:
        type: string
      node_type:
        type: string
      path:
        type: string
      permission_id:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.CommonMessage:
    properties:
      data:
        $ref: '#/definitions/structpb.Struct'
      table_slug:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.CreatePermissionRequest:
    properties:
      client_platform_id:
        type: string
      name:
        type: string
      node_type:
        type: string
      parent_id:
        type: string
      resource_type:
        type: integer
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetPermissionByIDResponse:
    properties:
      client_platform_id:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      permission_scopes:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.PermissionScope'
        type: array
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetPermissionListResponse:
    properties:
      count:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Permission'
        type: array
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetRolesResponse:
    properties:
      roles:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Role'
        type: array
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.PermissionScope:
    properties:
      client_platform_id:
        type: string
      method:
        type: string
      path:
        type: string
      permission_id:
        type: string
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.PermissionScopePrimaryKey:
    properties:
      client_platform_id:
        type: string
      method:
        type: string
      node_type:
        type: string
      path:
        type: string
      permission_id:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.UpdatePermissionRequest:
    properties:
      client_platform_id:
        type: string
      id:
        type: string
      name:
        type: string
      node_type:
        type: string
      parent_id:
        type: string
      resource_type:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://u-code.io
paths:
...
  /permission:
    get:
      consumes:
      - application/json
      description: Get Permission List
      operationId: get_permission_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetPermissionListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetPermissionListResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Permission List
      tags:
      - Permission
    post:
      consumes:
      - application/json
      description: Create Permission
      operationId: create_permission
      parameters:
      - description: CreatePermissionRequestBody
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Permission data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetPermissionByIDResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Permission
      tags:
      - Permission
    put:
      consumes:
      - application/json
      description: Update Permission
      operationId: update_permission
      parameters:
      - description: UpdatePermissionRequestBody
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permission data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetPermissionByIDResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Permission
      tags:
      - Permission
  /permission-scope:
    delete:
      consumes:
      - application/json
      description: Get PermissionScope
      operationId: delete_permission_scope
      parameters:
      - description: PermissionScopePrimaryKeyBody
        in: body
        name: permission-scope
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.PermissionScopePrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.CommonMessage'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete PermissionScope
      tags:
      - PermissionScope
    post:
      consumes:
      - application/json
      description: Create PermissionScope
      operationId: add_permission_scope
      parameters:
      - description: AddPermissionScopeRequestBody
        in: body
        name: permission-scope
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.AddPermissionScopeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: PermissionScope data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.PermissionScope'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create PermissionScope
      tags:
      - PermissionScope
  /permission/{permission-id}:
    delete:
      consumes:
      - application/json
      description: Get Permission
      operationId: delete_permission
      parameters:
      - description: permission-id
        in: path
        name: permission-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Permission
      tags:
      - Permission
    get:
      consumes:
      - application/json
      description: Get Permission By ID
      operationId: get_permission_by_id
      parameters:
      - description: permission-id
        in: path
        name: permission-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: PermissionBody
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetPermissionByIDResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Permission By ID
      tags:
      - Permission
...
  /role:
    get:
      consumes:
      - application/json
      description: Get Roles List
      operationId: get_roles_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: client-platform-id
        in: query
        name: client-platform-id
        type: string
      - description: client-type-id
        in: query
        name: client-type-id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetRolesListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetRolesResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Roles List
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Create Role
      operationId: create_role
      parameters:
      - description: AddRoleRequestBody
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.AddRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Role
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Update Role
      operationId: update_role
      parameters:
      - description: UpdateRoleRequestBody
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Role
      tags:
      - Role
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"